Issue #28 work. Implemented v2 -> v3 condition conversion.
```&&``` and ```||``` are converted to ```and``` and ```or```
respectively. ```VAR()``` functions have the condition used when they
are set substituted in their place, with the ```IF```s and ```IFNOT```s
switched if the condition is ```IF VAR()```. Language conditions are
removed from condition strings and are instead used to set the message
content language. The only trouble is that filenames are lowercased,
which isn't really an issue since Windows is case-insensitive.