Issue #79 work.
This should complete it, I can't think of any other plugin lists that
need colours added. The plugin list in the Editor now displays plugins
with accompanying BSAs in light blue. The grey text for empty plugins is
not used here because record parsing is skipped for the Editor, so they
all appear empty.