New conflicts filter UI (#224).
The option is now stored in each plugin's menu. Activating it for one
plugin prevents it being activated for others until it has been
deactivated again. Also improved the filtering so that a plugin with no
conflicts will cause all others to be hidden instead of no others. The
current filter target is also highlighted in the plugin list.

Finally, some minor attribute modification and semantics changes.