Fixed <loot-dialog>.
The issues seem to have been due to the layered attribute being true by
default in Polymer 0.5.0. The layering complicates matters (it moves
dialogs into shadow DOM), and I don't think I need it, so it's been
disabled. I also removed the open property, as it conflicts with an
existing function name.