Fixed docs styling in Firefox.
Since it doesn't support the Shadow DOM, the /deep/ CSS selector was
confusing it.

Also, I found that Firefox will only load fonts from a relative
(file://) URL if it's a subdirectory, so it won't load the bundled
Roboto font. It doesn't make sense to move the docs or the font files
though, so I guess Firefox users will just have to put up with Arial.