Added sidebar indication of open editors.
A pencil icon with a "Editor Is Open" tooltip will be displayed for any
plugins with open editors. The icon replaces the user metadata icon
while the editor is open, if the latter is visible.

Despite all the churn in the l10n template, the only real change is the
addition of the new tooltip string.