Try to fix repository parent path symlink issue.
For issue #417. Apparently if there is a symlink somewhere in the
checkout location's parent path, libgit2 will try to remove that
symlink. Searching the code for the error message given points to the
GIT_CHECKOUT_DONT_REMOVE_EXISTING flag controlling this behaviour,
though it seems to be intended only for paths within the checkout
directory.