Fix startup crash due to filesystem permissions.
An exception can be thrown when checking if a path exists, if the user
doesn't have permission to read that path. This exception wasn't getting
caught, so game detection was causing crashes on improperly-configured
systems. Now exceptions thrown during game detection are treated as the
game not being installed. Fixes #416.