Refactored git code.
The masterlist-specific code in has been moved into masterlist.cpp, and
the more general git-related code has been moved into a GitHelper class.

Although it's a helper, there are a few loot::error exceptions thrown in
its code, as it provides an error code that distinguishes git related
errors from others, so is useful.

There is still a lot of git code in masterlist.cpp. Not all of this will
be moved, but there may be some bits that can be.