Define a consistent graph vertex storage order.
The multithreading patch in 16ac349e8b48c901d97e28d78e34f0f61af0fef3
changed the order in which plugins are inserted into the
unordered map that is iterated over to create graph vertices.

With MSVC 2013, this insertion order doesn't seem to matter, and the
iteration order seems to be consistently the lexicographical order
of elements. With GCC 5.0 though, the insertion order does seem to
matter, as the multithreading patch's changes produced a different
iteration order.

The order in which graph vertices are created (and so stored, and
so iterated over) matters when tie-break edges are added, as earlier
tie-breaks can make potential later tie-breaks instead cause a cycle
and so be skipped, leading to a possibly different sorting order.

To resolve this issue, vertices are now created in lexicographical
order of their lowercased plugin filenames, in order to mimic the
previous apparent behaviour with MSVC 2013.