Rewrite installer using Inno Setup instead of NSIS.
For #467.

Korean and Simplified Chinese expect the unofficial translation
files to be dropped into the folder that holds the rest of the
translation files. If the files are missing, their translations
will not be included when the installer is built.

The translated strings were done in Google Translate and checked
in Bing Translate. They should at least be legible, if not great
quality.

The README now refers to Inno Setup, and the CONTRIBUTING file
has been generally updated both for the switch to Inno and the
recent code refactoring.