Deploy metadata validator on tagged builds.
Travis will deploy the metadata validator to a GitHub release for
any tagged build that passes the tests. If a release does not
already exist, it will be created, otherwise the file will be
added to the existing release.

The masterlist repositories can then use the deployed binary
instead of having to build the binary from source themselves on
every Travis build.

The api_key entry was generated by the Travis CLI tool using my
GitHub account details.