Remove game CRC cache
Non-plugin CRCs that were stored in the game's cache don't need to
be, as they are only used in conditions, which are cached
separately.

Plugin objects are now used to cache their CRC values along with
almost all plugin data.

This does remove the ability to cache plugin CRCs calculated when
evaluating conditions, but I think they were never used.