Overhaul version string extraction
* Use only one regex to match versions
* Don't store anything if there is no regex match

The new regex has been tested against all the description fields for
the plugins in the masterlists that have version conditions, and all
the plugins that I have installed.

However, without wider testing it is unknown whether it produces a
greater or fewer number of false positives than the previous regexes,
though as it is simpler and more clearly commented it's easier to fix
any that are found.

It's also worth noting that although this will extract the timestamps
found in Bashed Patch description fields, Pseudosem will only compare
the first number as it does not understand forwardslash separators.
This is not an issue though as the timestamp comparisons are not used
and are not useful, it's just easier to treat them as versions than
ignore them.

For #510.