Remove use of CEF sandbox
The sandbox library file is distributed prebuilt using MSVC 2013, which
means it can't be linked from MSVC 2015, which is in turn needed for
libloadorder's incoming thread safety support, as MSVC 2013 doesn't
support the thread_local keyword.

Removal of the sandbox shouldn't be an issue, as no external
resources/HTTP requests/etc. are used.