Handle priority values more sensibly.
While the YAML metadata syntax remains unchanged, the modulus of
the priority value is now stored by the PluginMetadata class
separately from whether or not the priority is global.

PluginMetadata.Priority() now returns this modulus priority value,
and new methods have been added to get and set whether the
priority is global or not.

The max_priority constant has also been renamed as it's actually
one greater than the maximum allowed priority value.