Use ESLint and the AirBnB style guide
To help enforce a consistent JS code style and improve code quality by
catching various syntax gotchas.

I'm editing in Atom with the linter-eslint package installed, and it
gives me inline warnings and errors, very handy. I'm not going to run
ESLint on Travis (yet) though, as it spits out far too many errors for
it to be useful.