Rewrite filters.js, and related plugin.js refactoring
filters.js now only exports a class that can be used to track filter
state, and contains functions for filtering using
Array.prototype.filter().

PluginCardContent is a new class for the data that gets displayed on
the front of plugin cards. It uses the filters to decide what is
visible and also converts the message data structure into something
a little simpler.

Plugin.modPriority and Plugin.isGlobalPriority have been renamed to
Plugin.priority and Plugin.isPriorityGlobal as they are more
sensible names.

I've also moved the filter state memory functions out of filters.js
as they're more settings-related. I've also moved the conflicting
plugins CEF callback and the filter application function out,
because they were to complex/tightly coupled to other components
to stay in there.

I expect this breaks code that Plugin and the filters are used in.

* The filters.hiddenPluginNo and filters.hiddenMessageNo have not
  yet been re-implemented and are currently missing. Given the
  complexity of the filtering, I'm thinking to just calculate
  their values at the end of a filter's event handler rather than
  tracking their state.
* Since filters.js now exports a class, loot.filters is no longer
  initialised,
* Getting conflicting plugins is also broken, as it should now be
  done in the conflict filter event handler, but that hasn't been
  updated.