Reimplement C++ LOOT settings handling
It deserves its own class. Improvements over the previous implementation
include:

* Unit tests
* If a setting is missing in loaded data, its default value is used
* Invalid game settings are skipped
* Settings are stored in named member variables instead of as untyped
map elements, so it's clear what settings exist and what their types
are.

Thread safety is ensured using a reader/writer mutex lock. A
readers/writer lock would be better, but isn't available in C++11 (it's
in C++17). A lock may not be necessary at all, as most settings changes
are called from JavaScript callbacks, which are single-threaded, but
I'm not sure how V8 actually runs them, so better to be safe than sorry.