Don't install settings file
The installer only installs a settings file so that it can set LOOT's
language to that chosen for the installer. However, as of
6307a3c1944d6b66a5cf3e666e985629c1e58553 LOOT can load partial settings
files, so the installer just needs to write out the language string to a
new settings file if one is not already present. This then means the
repository's settings.yaml can be removed.

Closes #527.