Replace LootState::UpdateGamesFromSettings()
With an load() that calls LootSettings::load() and enables/disables
debug logging and includes the functionality of
UpdateGamesFromSettings().

There isn't a test for this because it can only be tested for games
detected as installed, and none can be without significant changes to
the testing files/folders that would break most other tests.