Warn if load order has not been sorted since opening LOOT.
The warning is removed when the load order is sorted, but
reinstated if sorting is cancelled. Closes #528.

Note that if sorting is applied, then changes made so that
another sort produces a new load order and this is cancelled,
the message is not reinstated again, as its purposes is to remind
users that simply launching LOOT is not enough, and not to tell
them when another sort may be required.