Fix "You have not sorted..." reappearing
The "You have not sorted..." message was reappearing after sorting your
load order when no changes were necessary, then switching to another
game then back again, because the cancelSort query was reinstating the
message.

A new discardUnappliedChanges query was added to ensure the change
counter is zeroed, and that used where appropriate instead of cancelSort.