Add a splitter to the top of the editor panel
The splitter allows the height of the editor to be adjusted by dragging
it up and down. The height defaults to the height of the 'main' tab
elements, and is reset to this when the editor is closed. The height
of the editor cannot be greater than the available space, and
cannot be less than 56px.