Use CMake to resolve most dependencies
The CEF dependency is not resolved using CMake because it's only
available behind a CAPTCHA, and the Boost dependency is not resolved
using CMake because it is a relatively large dependency that is better
shared between LOOT and its dependencies that use it, instead of each
having to build a separate copy.

This is slower than the existing Travis script, because all targets in
the dependencies are now built, but it is more portable.