Remove Language::any
This breaks the API (again), but Language::any is a throwback from
before English-language strings were required for messages, and is not
useful now that they are.

Language::any picked the first message string, ignoring language, making
it unpredictable compared to any of the other language options, which
will provide the requested string or fall back to English if one is not
available.