Update to CEF 3.2704.1427.g95055fe
This CEF is based on Chromium 51, which drops support for Windows
Vista.

CEF now has builds hosted by Spotify that can be downloaded
automatically, so the build process has been updated to automate
downloading and building the CEF dependency. The version of CMake
in Ubuntu 12.04 doesn't support the URL_HASH parameter, so use a
more recent version from a PPA.

Unfortunately the minimal build's CMake config is broken
(https://bitbucket.org/chromiumembedded/cef/issues/1921) so for now
LOOT downloads the standard build (which is a much larger download).