Update to Polymer 1.6.0
It adds native CSS custom properties support, but this causes some
backwards compatibility / styling problems. To solve them, move
the custom properties out of the custom style element, and move
the custom  style element to be positioned after the elements it's
supposed to override (so in the <body>).

Also take advantage of native support for custom properties
throughout LOOT's external stylesheets.