Add Git pre-commit hook for generating POT file
It will generate a POT file, then add it to the commit if more than one
line is changed (because there's no point committing it if the only
change is to the creation time), and discard the generated file
otherwise.

The generated POT file changes the order of a few of its entries, so
I've included that in this commit.