Make the `git-commit` hook script POSIX compliant.
On a lot of systems using bashisms in /bin/sh scripts will be fine,
since /bin/sh is a symlink to /bin/bash, but some systems symlink it to
the more lightweight (and more strictly POSIX) /bin/dash or some other
shell. Safest thing is to not assume a bashism will work.

What has been changed:
1) `[[ … ]]` is a bashism. Replaced with `[ … ]`.
2) `-eq` is for numeric operations. When a variable is quoted, it
   becomes a string and uses `=` for comparison.
3) `&&` is a shell thing, but does not exist in `test`'s syntax.
   (`[ … ]` is shorthand for `test …`.) `-a` is the "and" when
   talking `test`.