Use the minimal CEF distribution
It's 1/2 or 1/5 (for Windows and Linux respectively) the size of the
standard distribution, compressed.

However, its CMakeLists.txt is an example for including into clients'
configs, and doesn't work with ExternalProject_Add, so ship a slightly
modified CMakeLists.txt to replace it (the only difference is the
cefsimple and cefclient include_directory lines have been removed).