Remove GameType::autodetect
It's never really necessary, the only place it was useful was in the
default GameSettings constructor, which is itself only used by the YAML
decoder (which requires a default constructor), and so using tes4 as a
default value instead is fine.

Removing it means a bunch of edge cases are removed, and API users can't
accidentally use something they're not supposed to.