Fix query error handling
Queries are executed in a separate thread from the query handler, and
exception handling needs to be done in that thread, so refactor
exception handling into Query::execute() and move query-specific logic
to Query::executeLogic().

Also simplify Query::executeLogic() to return the success string, since
they all throw on failure.