Resolve symlinks passed to CreateDatabase()
The API checks if the given game path and game local path are
directories, which isn't true if they are symlinks to directories, so if
they are symlinks, resolve them before performing the directory check.

Note that creating symlinks in Windows requires the
SeCreateSymbolicLinkPrivilege privilege, which only admins have by
default, so the test will fail unless run as admin or the privilege is
given to the user running the test.

Closes #645.