Refactor metadata class headers
This is unfinished, as there are a couple of classes that have
methods that take Game objects to eval metadata, and these will
probably not end up exposed in the API. None of the classes are
documented either.

YAML encode/decode code is now stored in separate headers that
depend on yaml-cpp and boost, and which will not be part of the
API proper.