Parse JS -> C++ requests using Protobuf
Protocol Buffers allows deserialisation code, including
objects that own the deserialised data, to be generated from
schema files, so that the deserialisation no longer depends
on the LOOT API's yaml-cpp implementation details. While
Protocol Buffers is a lot more complex than necessary for
parsing JSON, it provides the opportunity to later simplify
and switch to using the binary encoding (though this
complicates usage in JS), and to potentially reuse the same
schemas in the LOOT API, abstracting out the implementation to
some degree.

Protocol Buffers is treated as an external dependency (like
Boost) because it takes a while to build, it's easier to cache
on CI this way.

This also replaces the YAML output of Copy Content with JSON,
which is a relatively minor change as the feature is more of a
dump of what's in the UI than anything that's meant to be truly
human-readable.

The #undef ERROR lines are to avoid macro name collision
between CEF and Protobuf. It's also necessary to link to the
C++ runtime statically on AppVeyor, as Protobuf links
statically and the others don't seem to care.

This also reverts commit 414e830c2e5c9ec78a2b2426dc83d53622cee561.
The CMake Travis includes by default doesn't detect Protobuf
correctly, so go back to using a more recent release of CMake.

The added pthread linkage on Linux is necessary to have builds
pass on Travis, though it's unclear why.