Make games' local data paths configurable
Adds a local_path config variable for games. This is like the
--game-appdata-path CLI parameter but on a per-game basis. The
default value of an empty string means libloadorder (via the LOOT
API) will determine the path.

Add a column to the GUI game settings table so local_path values
are preserved, but hide it because the table is already too full.