Provide full path to plugins.txt in sorting error message
It's not the actual path, because LOOT doesn't know it. Instead it's the
path assuming that LOOT's folder for the game and the game's
own folder have the same name (which they do for all currently-supported
games). There are edge cases where this may be wrong, e.g. Oblivion
allows plugins.txt to be in the game's install directory.

Given that it's just a suggestion and plugins.txt may not be the actual
cause, it's not worth exposing this path through libloadorder and the
LOOT API.