Refactor recording window position on close
If getting the window position fails, don't prevent other settings from
getting saved.

Getting the position of a maximised window is a bit messy, because it
needs the window to be un-maximised first. It's then re-maximised in
case closing fails, so that the window isn't left in a different state.
This causes the window to visibly flash between the two states though.