Convert updateExists.js to TypeScript
The octokit.gitdata field is an alias of octokit.git, and the former
isn't in the current Octokit docs. It probably got deprecated at some
point, so I've replaced it with octokit.git.

The octokit.hasNextPage and octokit.getNextPage functions are in a
similar position, so have been replaced by octokit.paginate.