Convert translator.js to TypeScript
Disable the no-useless-constructor ESLint rule because it causes
parsing errors in type declaration files (where constructors have
no bodies).

It took a lot of trial and error to get Jest to accept the type
declaration file for Jed, I'm not sure it's right, but it seems to
work in Jest and in LOOT itself.