Update JS dependencies to resolve npm security advisory
minimist and mkdirp have been updated. Jest still has a
nested dependency on a vulnerable version of minimist, but there's
nothing I can do about that (and it's a dev-only issue).

I've also replaced the extract-zip dependency that I only just
switched to, because it depends on a specific vulnerable version of
mkdirp.