Replace usage of Inno Download Plugin
It's gone offline as Bitbucket has dropped support for Mercurial
repositories, which IDP used. Instead, include the MSVC redist in the
installer - this increases its size by 14 MB, but there's not really a
good alternative right now. However, Inno Setup 6.1 will include
built-in support for downloading files, so when that's released LOOT's
installer can switch to using that.