Add comments for translators to some text
babel-gettext-translator needed an update as it was writing the comments
as translator comments rather than extracted comments, which are treated
differently by tooling.

babel-gettext-translator requires the translator comment to immediately
precede the translation function call so that it can associate the
comment with the text. This leads to things like comments in the middle
of an assigment statement.

Comments for translators are intentionally not wrapped to 80 characters,
as line breaks are preserved when displayed by Poedit.