Support 64-bit registry view keys
On 64-bit Windows, Steam writes its install path keys in the 64-bit
view. To avoid complicating the UX for specifying keys, try the 32-bit
view first, and then try the 64-bit view if that fails. The 32-bit view
is tried first because most of the relevant keys are found there.