Fix installer for 64-bit LOOT builds
The installer would previously always look for and install the 32-bit
MSVC redistributable, even if the LOOT executable was 64-bit.

Building the installer for 64-bit LOOT requires the -DMyAppIs64Bit
parameter to be passed to Inno as it can't know what it's going to build
for ahead of time. The build_installer script exposes a -LOOTIs64Bit
switch for this.