Add support for Qt 5
This is useful because Qt 6 only supports 64-bit Windows 10 and
newer, which means that existing LOOT users on 32-bit versions of
Windows and Windows 7, 8 and 8.1 would not be able to update to a
Qt-based LOOT release.

The Steam survey statistics show that 4.12% of Steam's Windows
users run a 64-bit version of Windows 7, 8 or 8.1 and 0.35% run a
32-bit version of Windows. Although I think there are good
arguments for why it's reasonable to drop support for them, it's
better if we can avoid doing that without incurring too much
technical debt. Fortunately, this commit shows that the increase
in complexity required is pretty minimal.