Move LOOT game folders into a "games" subfolder
Having the game folders in the root of the LOOT data directory is
potentially problematic as a user could try to create a game that has
the same name as a LOOT data file or folder, so
segregate them in their own subdirectory.

Initialising a game will now cause LOOT to check for the game folder in
LOOT data folder and move it into the games subdirectory if it is found,
so game data is automatically migrated.