Add separate zlib dependency
zlib is required for minizip to support deflate compression.
minizip can download zlib itself at build time, but it does so using Git
and that tends to fail more often than not during local deveopment. In
doing so it also messes with what branch I'm on, which risks losing
commits.

To avoid all that, download and build zlib as a separate step and point
minizip to that zlib. zlib is built in source as otherwise when including
it minizip gets confused between its own CMake config and zlib's and the
build fails.

This may also be slightly faster, as the independent zlib build uses a
tarfile of a specific release instead of cloning the zlib repository
with all its history.