Support non-blocking background task execution
Query objects block the thread when they execute, but it'll be useful to
be able to run code that uses signals and slots (like checking for
LOOT updates does) and so doesn't block the thread it's executing on.

As such, introduce a Task abstraction that allows such code to run in
series on a background worker thread.