Add migration of repository URL and branch settings to sources
The migration can't be lossless, but it should either fail or succeed
outright and not result in mangled config. It can handle the following
cases:

- Any repository on GitHub. Old default repositories (for Skyrim VR and
  Fallout 4 VR) and old default branches will be updated to their latest
  defaults as they were before the move away from Git-based updates.
- Local file paths, so long as the path referenced is a Git repository
  with the relevant metadata file in the repository working copy's root.
  A warning will be logged if the repository doesn't have the configured
  branch currently checked out.

LOOT will log details if it's unable to migrate settings. Migration won't
happen if there's already config for the source alongside the old config
properties.