Replace CRLF with LF before calculating file Git blob hashes
The masterlists are committed with LF line endings, and that's what
they're downloaded from GitHub with, but if they come from a local Git
repository they may be checked out with Git's autocrlf feature enabled,
which replaces the LF with CRLF. If autocrlf is enabled, Git also
reverses the transformation before calculating the blob hash so that the
output of `git hash-object` matches what gets committed.

To ensure that the hashes LOOT calculates are consistent whether they're
downloaded or copied from a local Git repo's working directory, replace
CRLF with LF when loading from files.

This doesn't check if autocrlf is enabled, so will produce incorrect
hashes for a repository that's committed to with autocrlf disabled (i.e.
with CRLF line endings stored in Git), but that's relatively unlikely
and never the case for the official LOOT repos, and trying to find all
the relevant Git config files that may exist to check if autocrlf is in
use isn't worth the trouble.