Add ability to check for and fix ambiguous load orders
This replaces LOOT silently setting the load order even when it hadn't changed when sorting Skyrim or later games.

The check is performed after LOOT refreshes its load order state, so it should always detect ambiguity, not just when a game is first loaded.

The fix ambiguous load order menu item starts out disabled, is enabled when ambiguity is detected, and is disabled once the fix has happened.