Add support for games installed from the Microsoft Store
Older versions of the Microsoft Store and Xbox app install games to
the WindowsApps and ModifiableWindowsApps directories with very
restricted permissions, and newer versions install games to a
customisable location that defaults to <drive>:\XboxGames and doesn't
create any Registry entries for installed games.

Some Microsoft Store games are installed with each localisation in a
separate subdirectory, with no way for LOOT to know which the user
actually wants.

Some Microsoft Store games also use different local appdata paths to
store their load order data. With older installs, it looks like at
least some games can use one of two directories, and the logic for
which one is chosen is not well understood, and unfortunately it's
no longer possible to replicate these older installs.

All this makes compatibility a bit tricky, so the documentation now
details the issues.

Adding this support involved so much new code that it was worth
refactoring game detection into a separate source file.