Remove non-default timeout from tasks tests
The tests occasionally fails on GitHub Actions because tasks haven't
completed after 20 ms, so just use the default timeout of 5000 ms when a
task is expected to complete.

This means there remains a potential false positive when waiting for a
task that is expected to not complete, if it actually just takes longer
than 20 ms to complete, but that's acceptable.