Display notification dialog if sorting makes no changes
There may be a significant period of time between the user starting the
sorting process and it completing, enough that the user's attention
could leave the LOOT window and so they could miss the status bar
notification text saying that sorting made no changes and wonder what
happened.

It's not enough to just increase the timeout for the status bar text, as
it'll get replaced if anything else sets status text and so could still
be missed.

Instead, display the "no changes" message in a dialog box, so that the
user must interact with it before continuing to use LOOT. This means
the no-changes flow requires the same number of clicks as the
made-changes flow.