Save node positions to a binary file
Instead of using TOML, as the file is not meant to be human-readable.

This throws if any group name is longer than 64 KiB, so that when
reading bad input data from disk it can't end up causing huge memory
allocations (e.g. 2^64 bytes). This length restriction should be fine
as 64 KiB is still a ridiculously enormous group name.