Fix incorrect general info card sizes being calculated
The correct size was actually being calculated, the card's content just
didn't match what was expected for a given cache key. This can happen
because every general info card cache entry shared the same widget, but
the widget's content was only set when adding a new cache entry, so if
you went back to an older entry it would calculate the size using the
content set for the newest entry and give you the wrong result.