Fix card height calculation
A card's height depends on the width of the list's scrollable area, which
should equal the largest minimum width of all the cards in the list.
However, the list gets these widths by calling the delegate's
sizeHint() method, which calculates the card's width and height. This
results in circular logic, or the wrong height being calculated for that
first set of calls to sizeHint().

To avoid this, the minimum card widths for all plugin items are
calculated and cached when the items are added or changed, so that their
minimum widths are already known (and the largest can be found) by the
time the delegate's sizeHint() method runs.

As an optimisation, the widgets used to calculate these minimum widths
are cached instead of the widths themselves, so that the widgets can be
reused by the size hint calculation. Getting the minimum width from the
widget has a negligible cost.