Create sizing cards on demand if necessary
Sometimes when card content changes the delegate gets notified before
the card sizing cache, so no card exists for that content when the
sizeHint() method runs. This seems to be timing dependent as it doesn't
always happen.

As such, create the card if it doesn't exist when sizeHint() runs,
because sizeHint cannot fail and a card is necessary to succeed. This
might cause the size calculation for cards to be incorrect if the widest
card doesn't exist yet, so log that it happened.

In practice, the widest card is usually the general info card and it's
usually processed first, so it's unlikely to cause issues. As such, the
log severity is only info to avoid cluttering up the log when debug
logging is disabled.